+++ tests/unit_tests/Models/DomainTest.php
@@ -2,6 +2,8 @@
 
 namespace Pantheon\Terminus\UnitTests\Models;
 
+use League\Container\Container;
+use Pantheon\Terminus\Collections\DNSRecords;
 use Pantheon\Terminus\Collections\Domains;
 use Pantheon\Terminus\Collections\Workflows;
 use Pantheon\Terminus\Models\Environment;
@@ -17,9 +19,25 @@ use Pantheon\Terminus\Models\Workflow;
 class DomainTest extends ModelTestCase
 {
     /**
+     * @var Container
+     */
+    protected $container;
+    /**
+     * @var Environment
+     */
+    protected $environment;
+    /**
      * @var Domain
      */
     protected $model;
+    /**
+     * @var Workflow
+     */
+    protected $workflow;
+    /**
+     * @var Workflows
+     */
+    protected $workflows;
 
     public function setUp()
     {
@@ -31,6 +49,9 @@ class DomainTest extends ModelTestCase
     protected function _createDomain($attr)
     {
         $collection = $this->getMockBuilder(Domains::class)
+            ->disableOriginalConstructor()
+            ->getMock();
+        $this->container = $this->getMockBuilder(Container::class)
             ->disableOriginalConstructor()
             ->getMock();
         $this->workflow = $this->getMockBuilder(Workflow::class)
@@ -53,7 +74,7 @@ class DomainTest extends ModelTestCase
         $collection->method('getEnvironment')->willReturn($this->environment);
 
         $domain = new Domain((object)$attr, ['collection' => $collection,]);
-
+        $domain->setContainer($this->container);
         $domain->setRequest($this->request);
         return $domain;
     }
@@ -64,6 +85,42 @@ class DomainTest extends ModelTestCase
             ->method('request')
             ->with("sites/site id/environments/{$this->environment->id}/domains/dev.example.com", ['method' => 'delete',]);
         $this->model->delete();
+    }
+
+    /**
+     * Tests the Domain::getDNSRecords() function
+     */
+    public function testGetDNSRecords()
+    {
+        $data = (object)[
+            'id' => 'domain.com',
+            'dns_status_details' => (object)[
+                'dns_records' => [
+                    (object)[
+                        'type' => 'platform',
+                        'id' => 'live-mysite.pantheonsite.io',
+                        'status' => 'status',
+                        'status_message' => 'status message',
+                        'deletable' => false,
+                        'extraneous' => 'info',
+                    ]
+                ]
+            ]
+        ];
+        $domain = $this->_createDomain($data);
+
+        $records = $this->getMockBuilder(DNSRecords::class)
+            ->disableOriginalConstructor()
+            ->getMock();
+        $this->container->expects($this->once())
+            ->method('get')
+            ->with(
+                DNSRecords::class,
+                [['data' => $data->dns_status_details->dns_records,], ['domain' => $domain,],]
+            )
+            ->willReturn($records);
+
+        $this->assertEquals($records, $domain->getDNSRecords());
     }
     
     public function testSerialize()